+++ include/kf2/game/adaptive_control_client.hpp
@@ -17,7 +17,6 @@ enum class AdaptiveResourceControl : std::uint8_t {
     ram,
     mixed,
     recover,
-    target_fps,
 };
 
 struct AdaptiveControlRequest final {